refactor: move SpanMetricsInstrumentor into connector package
Relocate instrumentor.py from cloudwatch/ into cloudwatch/connector/
alongside span_metrics_connector.py, and update the entry point and
imports (__init__, tests) to the new module path.
